packages/fine-tuning/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
dependencies = [
  "datasets>=3.5.0",
  "fastapi[standard]",
  "hydra-core>=1.3.2",
  "ipywidgets>=8.1.6",
  "jupyter>=1.1.1",
  "numpy>=2.2.4",
  "rich>=14.0.0",
  "torch",
  "torchvision",
  "transformers>=4.51.3",
  "trl>=0.15.2",
  "unsloth>=2025.3.19",
  "triton==3.2.0; sys_platform == 'linux'",
  "xformers",
  "python-dotenv>=1.1.0",
  "boto3>=1.37.35"
]
description = "Fine tuning service for Resonare"
name = "fine-tuning"
requires-python = ">=3.12"
version = "0.1.0"

[tool.hatch.build.targets.wheel]
packages = ["src/fine-tuning"]

[[tool.uv.index]]
explicit = true
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"

[tool.uv.sources]
torch = [
  {index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'"}
]
torchvision = [
  {index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'"}
]
xformers = [
  {index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'"}
]
